+++ src/main/java/org/chatchat/chatroom/domain/ChatRoom.java
@@ -1,4 +1,4 @@
-package org.chatchat.entity;
+package org.chatchat.chatroom.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,6 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.chatchat.chatpart.domain.ChatPart;
+import org.chatchat.common.entity.BaseEntity;
 
 import java.util.List;
