+++ src/main/java/org/chatchat/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package org.chatchat.config;
+package org.chatchat.common.config;
 
 
 import org.springframework.context.annotation.Bean;
