+++ src/main/java/org/chatchat/common/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package org.chatchat.exception;
+package org.chatchat.common.exception;
 
-import org.chatchat.exception.type.ErrorType;
+import org.chatchat.common.exception.type.ErrorType;
 import org.springframework.http.HttpStatus;
 
 public class NotFoundException extends ApiException {
