+++ src/main/java/org/chatchat/user/domain/User.java
@@ -1,4 +1,4 @@
-package org.chatchat.entity;
+package org.chatchat.user.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,6 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.chatchat.chatpart.domain.ChatPart;
+import org.chatchat.common.entity.BaseEntity;
 
 import java.util.List;
 
@@ -21,6 +23,9 @@ public class User extends BaseEntity {
 
     @Column(nullable = false, unique = true)
     private String email;
+
+    @Column(nullable = false, unique = true)
+    private String name;
 
     @Column(nullable = false)
     private String password;
