+++ src/main/java/org/chatchat/user/domain/UserRepository.java
@@ -1,6 +1,6 @@
-package org.chatchat.repository;
+package org.chatchat.user.domain;
 
-import org.chatchat.entity.User;
+import org.chatchat.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
